tests/requirements.txt
# Core dependencies
transformers>=4.30.0
pillow>=10.0.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.65.0
pyarrow>=14.0.1
mercantile>=1.2.1
owslib>=0.31.0
mapbox-vector-tile>=2.0.1
shapely>=2.0.0,<3.0.0
pyproj>=3.6.1
pystac>=1.8.0
redis>=5.0.0
nltk>=3.8.1
diffusers>=0.25.0
langchain>=0.1.0
langchain-community>=0.0.1
duckdb>=0.9.0
geopy>=2.4.1
folium>=0.15.1
rtree>=1.1.0
aiohttp>=3.9.0
fsspec>=2024.2.0
cryptography>=42.0.0
pyjwt>=2.8.0
pystac-client>=0.8.3
planetary-computer>=1.0.0
fastapi>=0.109.0
netCDF4>=1.6.5
earthengine-api>=0.1.390
typing-extensions>=4.9.0
pydantic>=2.6.0
uvicorn>=0.27.0
python-multipart>=0.0.9
osmnx>=1.9.0
py6s>=1.9.0
opencv-python>=4.8.0
noise>=1.2.2

# Scientific computing dependencies
numpy>=1.24.4,<2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
ipywidgets>=8.0.0
scikit-learn>=1.3.0
rasterio>=1.3.8
geopandas>=0.14.0
albumentations>=1.3.1
sentence-transformers>=2.2.0
xarray>=2023.0.0
dask>=2024.1.0
accelerate>=1.3.0
scipy>=1.11.0
scipy-openblas64>=0.3.23
faiss-cpu>=1.7.4

# Test dependencies
pytest>=8.3.4
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
pytest-benchmark>=4.0.0
pytest-timeout>=2.2.0

# Development dependencies
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
pre-commit>=3.6.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
sphinx-autodoc-typehints>=2.0.0
nbsphinx>=0.9.0
sphinx-copybutton>=0.5.2
myst-parser>=0.18.1
pandoc>=2.3

# Cloud Provider SDKs
google-cloud-core>=2.3.3
google-api-core>=2.11.1
google-cloud-storage>=2.10.0
google-cloud-monitoring>=2.15.1
boto3>=1.34.11
azure-mgmt-compute>=30.3.0
azure-mgmt-network>=25.1.0 